Updated the code for writing the timestamp
Now the code counts the number of bytes to the write point on its own
(meaning that edits can be made to the file without the possibility of
breaking it, so long as these edits do not affect the line that we are
writing to).

The executable name can also be anything now, and the program will
automatically strip that name off of the path regardless of what it is.